Midway converting iterative tests
